+++ Assets/Scripts/SetStatsInput.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a06628d4d0c84a88a2aa7e5af9a96ae5
+timeCreated: 1732119278
